@@ -1,11 +1,11 @@
-﻿using System;
-namespace MyBlog.BL.General
-{
-    public class Constants
-    {
-        public const string RememberMeCookieName = "RememberMe";
+﻿namespace MyBlog.BL.General;
 
-        public const string SessionCookieName = "CustomSessionId";
-    }
+public class Constants
+{
+    public const string RememberMeCookieName = "RememberMe";
+
+    public const string SessionCookieName = "CustomSessionId";
+
+    public const int RememberMeDays = 30;
 }
 
