@@ -24,7 +24,7 @@ public class WeatherForecastController : ControllerBase
     [HttpGet(Name = "GetWeatherForecast")]
     public IEnumerable<WeatherForecast> Get()
     {
-        string? user = User.Identity.Name;
+        string? user = User.Identity?.Name;
 
         return Enumerable.Range(1, 5).Select(index => new WeatherForecast
         {
