@@ -1,6 +1,4 @@
-﻿using Microsoft.AspNetCore.DataProtection;
-using Microsoft.AspNetCore.Http;
-using Microsoft.EntityFrameworkCore.Metadata.Internal;
+﻿using Microsoft.VisualBasic;
 using MyBlog.DAL.Interfaces;
 using MyBlog.DAL.Models;
 
@@ -10,15 +8,24 @@ namespace MyBlog.BL.Auth
     {
         private readonly IEncrypt encrypt;
 
+        // очень плохо держать тут IHttpContextAccessor, его не должно быть в BL уровне
+        // но для простоты кода в книге я оставлю его здесь, а на бусти https://boosty.to/mflenov
+        // у меня есть видео создания электронного магазина и сайта на  .NET и там я сделал 
+        // рефакторинг для и код там лучше. Здесь в коде есть файл WebCookie.cs, который нужно использовать 
+        // вместо прямого доступа к httpContextAccessor
         private readonly IHttpContextAccessor httpContextAccessor;
         private readonly IAuthenticationDAL authenticationDAL;
         private readonly IAuthentication authentication;
         private readonly ISession session;
+        private readonly IUserTokenDAL userTokenDAL;
+        private readonly IWebCookie webCookie;
 
         public CurrentUser(IHttpContextAccessor httpContextAccessor,
             IAuthenticationDAL authenticationDAL,
             IAuthentication authentication,
             ISession session,
+            IUserTokenDAL userTokenDAL,
+            IWebCookie webCookie,
             IEncrypt encrypt)
         {
             this.httpContextAccessor = httpContextAccessor;
@@ -26,6 +33,8 @@ namespace MyBlog.BL.Auth
             this.authentication = authentication;
             this.encrypt = encrypt;
             this.session = session;
+            this.userTokenDAL = userTokenDAL;
+            this.webCookie = webCookie;
         }
 
         private UserModel? userData = null;
@@ -33,7 +42,8 @@ namespace MyBlog.BL.Auth
         {
             if (userData == null)
             {
-                if (IsLoggedIn())
+                bool isLoggedIn = await IsLoggedIn();
+                if (isLoggedIn)
                 {
                     userData = await authenticationDAL.GetUser(httpContextAccessor?.HttpContext?.Session.GetInt32("userid") ?? 0);
                     //userData = await authenticationDAL.GetUser((int)session.GetUserId().GetAwaiter().GetResult());
@@ -44,15 +54,39 @@ namespace MyBlog.BL.Auth
             return userData!;
         }
 
-        public bool IsLoggedIn()
+		public async Task<int?> GetUserIdByToken()
+		{
+			string? tokenCookie = webCookie.Get(General.Constants.RememberMeCookieName);
+			if (tokenCookie == null)
+				return null;
+			Guid? tokenGuid = General.Helpers.StringToGuidDef(tokenCookie ?? "");
+			if (tokenGuid == null)
+				return null;
+
+            int? userid = await userTokenDAL.Get((Guid)tokenGuid);
+			return userid;
+        }
+
+
+        public async Task<bool> IsLoggedIn()
         {
-            bool loggedIn = this.httpContextAccessor?.HttpContext?.Session.GetInt32("userid") != null;
+            //bool loggedIn = this.httpContextAccessor?.HttpContext?.Session.GetInt32("userid") != null;
 
             // не используйте GetAwaiter().GetResult(), я использую это здесь только потому
             // что не хочу рефакторить код и оставить его так, как я его приводил во второй главе
-            //bool loggedIn = session.IsLoggedIn().GetAwaiter().GetResult();
-            if (!loggedIn)
+            bool loggedIn = await session.IsLoggedIn();
+            if (!loggedIn) // Запомни меня
             {
+                // Это вариант с уникальными токенами
+				int? userid = await GetUserIdByToken();
+				if (userid != null)
+				{
+					await session.SetUserId((int)userid);
+					loggedIn = true;
+                }
+
+                /* 
+                // ЭТО ВАРИАНТ С шифрованием UserId. Он работает, но небезопасный
                 var cookie = httpContextAccessor?.HttpContext?.Request?.Cookies.FirstOrDefault(m => m.Key == General.Constants.RememberMeCookieName);
                 if (cookie != null && cookie.Value.Value != null)
                 {
@@ -74,9 +108,31 @@ namespace MyBlog.BL.Auth
                         return false;
                     }
                 }
+                */
             }
             return loggedIn;
-        }
+        }  
+
+		private Guid? GetCurrentUserToken() {
+			string? tokenCookie = webCookie.Get(General.Constants.RememberMeCookieName);
+			if (tokenCookie == null)
+				return null;
+			return General.Helpers.StringToGuidDef(tokenCookie ?? "");
+		}
+
+        public async Task Logout()
+		{
+			await session.DeleteSessionId();
+
+			Guid? tokenGuid = GetCurrentUserToken();
+			if (tokenGuid != null){
+				await userTokenDAL.DeleteToken((Guid)tokenGuid);
+				webCookie.Delete(General.Constants.RememberMeCookieName);
+			}
+
+            webCookie.Delete(General.Constants.SessionCookieName);
+            session.ResetSessionCache();
+        }    
     }
 }
 
