@@ -4,9 +4,11 @@ namespace MyBlog.BL.Auth
 {
     public interface ICurrentUser
     {
-        bool IsLoggedIn();
+        Task<bool> IsLoggedIn();
 
         Task<UserModel> GetUserData();
+
+        Task Logout();
     }
 }
 
