@@ -1,13 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using MyBlog.Models;
 using MyBlog.BL.Blog;
-using Microsoft.Extensions.DependencyInjection;
 using MyBlog.BL.Auth;
-using Microsoft.AspNetCore.Http;
 using System.Security.Cryptography;
 using MyBlog.Middleware;
 
@@ -27,16 +21,17 @@ namespace MyBlog.Controllers
 
         [HttpGet]
         [Route("/blog/add")]
-        public IActionResult AddAction()
+        public async Task<IActionResult> AddAction()
         {
-            if (currentUser.IsLoggedIn() == false)
+            bool isLoggedIn = await currentUser.IsLoggedIn();
+            if (isLoggedIn == false)
                 return Redirect("/Login");
             return View("Edit", new BlogViewModel());
         }
 
         [HttpPost]
         [Route("/blog/add")]
-        public async Task<IActionResult> AddPostAction(BlogViewModel model)
+        public async Task<IActionResult> AddPostAction([Bind("Title")]BlogViewModel model)
         {
             // Этот код можно отделить от контроллера,
             // но я оставлю здесь 
