@@ -1,6 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using Azure.Core;
+﻿
+using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
 using MyBlog.BL.Auth;
 using MyBlog.Models;
@@ -10,26 +9,31 @@ namespace MyBlog.Controllers
     public class LoginController : Controller
     {
         private readonly IAuthentication authentication;
+        private readonly IHttpContextAccessor httpContextAccessor;
 
-        public LoginController(IAuthentication authentication)
+        public LoginController(IAuthentication authentication, IHttpContextAccessor httpContextAccessor)
         {
             this.authentication = authentication;
+            this.httpContextAccessor = httpContextAccessor;
         }
 
         [HttpGet]
         [Route("/login")]
         public IActionResult Index(string u)
         {
+            ViewBag.YaUrl = YaAuthViewModel.GetAuthUrl(httpContextAccessor);
             return View(new LoginViewModel()
             {
                 U = u
             });
         }
 
-        [HttpPost]
+        [HttpPost]        
         [Route("/login")]
+        [EnableCors("Loginpolicytest")]
         public async Task<IActionResult> IndexPost(LoginViewModel model)
         {
+            ViewBag.YaUrl = YaAuthViewModel.GetAuthUrl(httpContextAccessor);
             if (ModelState.IsValid)
             {
                 bool isLocked = await authentication.IsAccountLocked(model.Email!);
