@@ -28,7 +28,7 @@ namespace MyBlog.DAL.Implementations.Dapper
             using (var connection = new SqlConnection(DbHelper.GetConnectionString()))
             {
                 await connection.OpenAsync();
-                string sql = @"select SessionID, Content, Created, LastAccessed, UserId from [Session] where sessionId = @sessionId";
+                string sql = @"select SessionID, Content, Created, LastAccessed, UserId  from [Session] where sessionId = @sessionId";
 
                 var sessions = await connection.QueryAsync<SessionModel>(sql, new { sessionId = sessionId });
                 return sessions.FirstOrDefault();
@@ -41,13 +41,35 @@ namespace MyBlog.DAL.Implementations.Dapper
             {
                 await connection.OpenAsync();
                 string sql = @"update [Session]
-                      set Content = @Content, LastAccessed = @LastAccessed, UserId = @UserId
+                      set Content = @Content
                       where SessionID = @SessionID
                 ";
 
                 return await connection.ExecuteAsync(sql, model);
             }
         }
+
+        public async Task Extend(Guid dbSessionID)
+        {
+            using (var connection = new SqlConnection(DbHelper.GetConnectionString()))
+            {
+                string sql = @"update [Session]
+                        set LastAccessed = @lastAccessed
+                        where SessionID = @dbSessionID";
+
+                await connection.ExecuteAsync(sql, new { dbSessionID = dbSessionID, lastAccessed = DateTime.Now });
+            }
+        }
+
+        public async Task Delete(Guid sessionId)
+        {
+            using (var connection = new SqlConnection(DbHelper.GetConnectionString()))
+            {
+                string sql = @"delete from [Session] where SessionID = @dbSessionID";
+
+                await connection.ExecuteAsync(sql, new { dbSessionID = sessionId });
+            }
+        }                
     }
 }
 
