@@ -9,7 +9,7 @@
 </div>
 <div class="register-form">
     <form method="post" enctype="multipart/form-data">
-        <input type="hidden" userid="@Model.Title" />
+        <input type="hidden" name="userid" value="1" />
         <p>
             <label>Заголовок</label>
             <input class="form-input" type="text" name="title" value="@Model.Title" />
