@@ -16,7 +16,7 @@
         </p>
         <p>
             <label>Пароль</label>
-            <input class="form-input" type="password" name="Password" value="" />
+            <input class="form-input" type="password" name="Password" value="" autocomplete="new-password" />
             <div class="error">@Html.ValidationMessageFor(m => m.Password)</div>
             <span class="hint">
                 Пароль должен:<br />
